fix: collect aws boundary tests
